Extract AbstractDbTimestampBoundStoreTest for Postgres and Oracle
